Added cleanup logic in case of instance non active

+++ builder/ibmcloud/vpc/step_create_instance.go
@@ -370,34 +370,36 @@ func (step *stepCreateInstance) Cleanup(state multistep.StateBag) {
 
 	// Delete Floating IP if it was created (VSI Interface was set as public)
 	if config.VSIInterface == "public" {
-		floatingIP := state.Get("floating_ip").(string)
-		ui.Say(fmt.Sprintf("Releasing the Floating IP: %s ...", floatingIP))
+		if state.Get("floating_ip") != nil {
+			floatingIP := state.Get("floating_ip").(string)
+			ui.Say(fmt.Sprintf("Releasing the Floating IP: %s ...", floatingIP))
 
-		floatingIPID := state.Get("floating_ip_id").(string)
+			floatingIPID := state.Get("floating_ip_id").(string)
 
-		options := vpcService.NewGetFloatingIPOptions(floatingIPID)
-		floatingIPresponse, _, err := vpcService.GetFloatingIP(options)
-		if err != nil {
-			err := fmt.Errorf("[ERROR] Error getting the Floating IP: %s", err)
-			state.Put("error", err)
-			ui.Error(err.Error())
-			// log.Fatalf(err.Error())
-			return
-		}
-		status := floatingIPresponse.Status
-		if *status == "available" {
-			options := vpcService.NewDeleteFloatingIPOptions(floatingIPID)
-			result, err := vpcService.DeleteFloatingIP(options)
-
+			options := vpcService.NewGetFloatingIPOptions(floatingIPID)
+			floatingIPresponse, _, err := vpcService.GetFloatingIP(options)
 			if err != nil {
-				err := fmt.Errorf("[ERROR] Error releasing the Floating IP. Please release it manually: %s", err)
+				err := fmt.Errorf("[ERROR] Error getting the Floating IP: %s", err)
 				state.Put("error", err)
 				ui.Error(err.Error())
 				// log.Fatalf(err.Error())
 				return
 			}
-			if result.StatusCode == 204 {
-				ui.Say("The Floating IP was successfully released!")
+			status := floatingIPresponse.Status
+			if *status == "available" {
+				options := vpcService.NewDeleteFloatingIPOptions(floatingIPID)
+				result, err := vpcService.DeleteFloatingIP(options)
+
+				if err != nil {
+					err := fmt.Errorf("[ERROR] Error releasing the Floating IP. Please release it manually: %s", err)
+					state.Put("error", err)
+					ui.Error(err.Error())
+					// log.Fatalf(err.Error())
+					return
+				}
+				if result.StatusCode == 204 {
+					ui.Say("The Floating IP was successfully released!")
+				}
 			}
 		}
 	}
@@ -440,24 +442,25 @@ func (step *stepCreateInstance) Cleanup(state multistep.StateBag) {
 	}
 
 	// Deleting Security Group's rule
+	if state.Get("security_group_rule_id") != nil {
+		ruleID := state.Get("security_group_rule_id").(string)
+		ui.Say(fmt.Sprintf("Deleting Security Group's rule %s ...", ruleID))
+		sgRuleOptions := &vpcv1.DeleteSecurityGroupRuleOptions{}
+		sgRuleOptions.SetSecurityGroupID(state.Get("security_group_id").(string))
+		sgRuleOptions.SetID(ruleID)
+		sgRuleResponse, sgRuleErr := vpcService.DeleteSecurityGroupRule(sgRuleOptions)
 
-	ruleID := state.Get("security_group_rule_id").(string)
-	ui.Say(fmt.Sprintf("Deleting Security Group's rule %s ...", ruleID))
-	sgRuleOptions := &vpcv1.DeleteSecurityGroupRuleOptions{}
-	sgRuleOptions.SetSecurityGroupID(state.Get("security_group_id").(string))
-	sgRuleOptions.SetID(ruleID)
-	sgRuleResponse, sgRuleErr := vpcService.DeleteSecurityGroupRule(sgRuleOptions)
+		if sgRuleErr != nil {
+			sgRuleErr := fmt.Errorf("[ERROR] Error deleting Security Group's rule %s. Please delete it manually: %s", ruleID, sgRuleErr)
+			state.Put("error", sgRuleErr)
+			ui.Error(err.Error())
+			// log.Fatalf(err.Error())
+			return
+		}
 
-	if sgRuleErr != nil {
-		sgRuleErr := fmt.Errorf("[ERROR] Error deleting Security Group's rule %s. Please delete it manually: %s", ruleID, sgRuleErr)
-		state.Put("error", sgRuleErr)
-		ui.Error(err.Error())
-		// log.Fatalf(err.Error())
-		return
-	}
-
-	if sgRuleResponse.StatusCode == 204 {
-		ui.Say("The Security Group's rule was successfully deleted!")
+		if sgRuleResponse.StatusCode == 204 {
+			ui.Say("The Security Group's rule was successfully deleted!")
+		}
 	}
 
 	// Wait a couple of seconds before attempting to delete the security group.
@@ -465,22 +468,24 @@ func (step *stepCreateInstance) Cleanup(state multistep.StateBag) {
 
 	// Deleting Security Group
 	if config.SecurityGroupID == "" {
-		securityGroupName := state.Get("security_group_name").(string)
-		ui.Say(fmt.Sprintf("Deleting Security Group %s ...", securityGroupName))
-		securityGroupID := state.Get("security_group_id").(string)
-		sgOptions := &vpcv1.DeleteSecurityGroupOptions{}
-		sgOptions.SetID(securityGroupID)
-		sgResponse, err := vpcService.DeleteSecurityGroup(sgOptions)
-		if err != nil {
-			err := fmt.Errorf("[ERROR] Error deleting Security Group %s. Please delete it manually: %s", securityGroupName, err)
-			state.Put("error", err)
-			ui.Error(err.Error())
-			// log.Fatalf(err.Error())
-			return
-		}
+		if state.Get("security_group_name") != nil {
+			securityGroupName := state.Get("security_group_name").(string)
+			ui.Say(fmt.Sprintf("Deleting Security Group %s ...", securityGroupName))
+			securityGroupID := state.Get("security_group_id").(string)
+			sgOptions := &vpcv1.DeleteSecurityGroupOptions{}
+			sgOptions.SetID(securityGroupID)
+			sgResponse, err := vpcService.DeleteSecurityGroup(sgOptions)
+			if err != nil {
+				err := fmt.Errorf("[ERROR] Error deleting Security Group %s. Please delete it manually: %s", securityGroupName, err)
+				state.Put("error", err)
+				ui.Error(err.Error())
+				// log.Fatalf(err.Error())
+				return
+			}
 
-		if sgResponse.StatusCode == 204 {
-			ui.Say("The Security Group was successfully deleted!")
+			if sgResponse.StatusCode == 204 {
+				ui.Say("The Security Group was successfully deleted!")
+			}
 		}
 	}
